SlideIntoScreenView is created
